+++ code/Editor/Pipeline/PipelineDependsIncremental.cpp
@@ -409,7 +409,7 @@ void PipelineDependsIncremental::updateDependencyHashes(
 			{
 				uint8_t buffer[4096];
 				Adler32 a32;
-				int32_t r;
+				int64_t r;
 
 				a32.begin();
 				while ((r = dataStream->read(buffer, sizeof(buffer))) > 0)
@@ -456,7 +456,7 @@ void PipelineDependsIncremental::updateDependencyHashes(
 		{
 			uint8_t buffer[4096];
 			Adler32 a32;
-			int32_t r;
+			int64_t r;
 
 			a32.begin();
 			while ((r = fileStream->read(buffer, sizeof(buffer))) > 0)
